requirements.txt
bokeh==2.4.3
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
torch==1.12.1+cu116
torchvision==0.13.1+cu116
